add inklist component
