+++ packages/nextjs/.eslintignore
@@ -9,3 +9,4 @@ node_modules/
 **/*.png
 **/*.svg
 **/generated/**/*
+types/*.ts
